Fix merge clash.

+++ bundles/org.openhab.binding.ipcamera/src/main/java/org/openhab/binding/ipcamera/config/CameraConfig.java
@@ -161,6 +161,6 @@ public class CameraConfig {
     }
 
     public int getPort() {
-        return port.intValue();
+        return port;
     }
 }
